Automatically backed up by Learn

+++ db/migrate/002_create_costume_stores.rb
@@ -5,7 +5,7 @@ class CreateCostumeStores < ActiveRecord::Migration[4.2]
     create_table :costume_stores do |t|
       t.string :name 
       t.text :location
-      t.float :costume_inventory 
+      t.integer :costume_inventory 
     end
   end
